# Conflicts:
#	source/extensions/filters/network/common/redis/codec_impl.cc
#	source/extensions/filters/network/common/redis/codec_impl.h
#	source/extensions/filters/network/common/redis/utility.cc
#	source/extensions/filters/network/redis_proxy/command_splitter_impl.cc
#	source/extensions/filters/network/redis_proxy/command_splitter_impl.h
#	test/extensions/filters/network/common/redis/codec_impl_test.cc
#	test/extensions/filters/network/redis_proxy/command_split_speed_test.cc
#	test/extensions/filters/network/redis_proxy/command_splitter_impl_test.cc

+++ source/extensions/filters/network/redis_proxy/conn_pool_impl.cc
@@ -267,7 +267,7 @@ Common::Redis::Client::PoolRequest* InstanceImpl::ThreadLocalPool::makeRequestTo
     host_address_map_key = host_address;
   } else {
     const auto ip_port = absl::string_view(host_address).substr(colon_pos + 1);
-    uint64_t ip_port_number;
+    uint32_t ip_port_number;
     if (!absl::SimpleAtoi(ip_port, &ip_port_number) || (ip_port_number > 65535)) {
       return nullptr;
     }
@@ -290,7 +290,7 @@ Common::Redis::Client::PoolRequest* InstanceImpl::ThreadLocalPool::makeRequestTo
     if (!ipv6) {
       // Only create an IPv4 address instance if we need a new Upstream::HostImpl.
       const auto ip_port = absl::string_view(host_address).substr(colon_pos + 1);
-      uint64_t ip_port_number;
+      uint32_t ip_port_number;
       if (!absl::SimpleAtoi(ip_port, &ip_port_number) || (ip_port_number > 65535)) {
         return nullptr;
       }
